Move transcript components to a separate files
